requirements.txt
numpy==1.24.3
pandas==2.0.1
sentence_transformers==2.2.2
streamlit==1.22.0
torch==2.0.1
trubrics==1.3.6
arxiv==1.4.7
python-dotenv==1.0.0
hopsworks==3.0.5